backend/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
Flask-MySQL==1.6.0
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
PyMySQL==1.1.2
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
